requirements.txt
# TPC-DS 流批一体处理 Demo 依赖包
# 作者：Vance Chen

# MySQL 数据库连接器
mysql-connector-python==8.2.0

# 数据处理和生成
faker==20.1.0

# 可选依赖（按需启用）
# pandas==2.1.4          # 数据分析
# numpy==1.24.4           # 数值计算
# matplotlib==3.7.2       # 数据可视化 